// src/app/models/flex-notifications.model.ts
export class Notificationmodel {
  Id: number;
  Name: string;
  Description: string;
  Step: string;
  Action: string;
  Template: string;
  Subject: string;
  Body: string;
  NotificationType: string;
  TemplateId: number;
  TemplateName: string;
  StepActionId: number;
  ScopeEntityType: string;
  ScopeEntityId: number;
  CreatedById: number;
  UpdatedById: number;
  EmailSubject: string;
  EmailBody: string;
  CreatedByRoleId: number;
  UpdatedByRoleId: number;
  ParameterCarrier: string;
}